Stamina, enemy AI and Ammo for pistol
Added in a stamina bar function, however it will deplete very fast and causes a error and won't regenerate so that will be fixed in the future. The enemy is on ground and follows player once they are in a field of view by navmesh. Finally, the pistol has a current ammo deplete, once it's gone, it stops shooting.