Scenes, Guns and Health
This one had the health for both enemy and player sorted out and they can be killed. Player gets killed and goes to game over screen. Win screen is implemented. The guns now have a line to know where they are shooting, have a fire rate and a ammo count. Finally, a start to swap weapons was implemented.