Keys and doors, multiple patrol enemies, win
There is multiple enemies that patrol a little area they are in, they still don't return to their original patrol once finished chasing the player but they do a little patrol where they are at. The doors have animations and open once key is picked up and pressed interact buttons. There is finally, a way to win the game and that is to get to the bed in the house.